pyproject.toml
[tool.poetry]
name = "rag-data-pipeline"
version = "0.1.0"
description = ""
authors = ["Prakash P. <47575786+prakash360@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pymupdf = "^1.24.4"
sentence-transformers = "^2.7.0"
chromadb = "^0.5.0"
langchain = "^0.2.0"
tiktoken = "^0.7.0"
langchain-community = "^0.2.0"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
pypdf = "^4.2.0"
langchain-openai = "^0.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
